This is a a new user design which emphasiszes more polymorphism and abstraction of user class
